backend/requirements.txt
dj-database-url==0.5.0
Django==2.2.12
django-extensions==2.2.9
djangorestframework==3.11.0
python-decouple==3.3
